Fixed data type size issue
sizeof(long) in clear_zbuff() should have been sizeof(int). This caused a crash just before the torus when running BB at resolutions >= 2880*1800 (possibly even a bit lower).

+++ tex.c
@@ -92,7 +92,7 @@ void unset_zbuff()
 
 static inline void clear_zbuff()
 {
-    memset(zbuff, 0x55, (X_s * Y_s * sizeof(long)));
+    memset(zbuff, 0x55, (X_s * Y_s * sizeof(int)));
 }
